requirements.txt
metadrive-simulator==0.2.4


attrs==21.4.0
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.1.0
commonmark==0.9.1
contextvars==2.4
cycler==0.11.0
# dataclasses==0.8
decorator==4.4.2
DI-engine==0.2.3
easydict==1.9
elementpath==2.4.0
enum-tools==0.9.0.post1
ephem==4.1.3
Flask==1.1.4
gym==0.20.0
h5py==3.1.0
idna==3.3
imageio==2.15.0
imgaug==0.4.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lmdb==1.3.0
loguru==0.3.0
lz4==3.1.10
MarkupSafe==2.0.1
matplotlib==3.3.4
namedlist==1.8
networkx==2.5.1
numpy==1.19.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Panda3D==1.10.11
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==1.1.5
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
py-trees==0.8.3
pycparser==2.21
pydot==1.4.2
pygame==2.1.2
Pygments==2.12.0
pynng==0.7.1
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==5.4.1
readerwriterlock==1.0.9
requests==2.27.1
responses==0.12.1
rich==12.0.1
scikit-image==0.17.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.10
tensorboardX==2.2
terminaltables==3.1.10
tifffile==2020.9.3
tomli==1.2.3
# torch==1.8.1+cu111
# torchaudio==0.8.1
# torchvision==0.9.1+cu111
tqdm==4.64.0
trueskill==0.4.5
typing_extensions==4.1.1
urllib3==1.26.10
URLObject==2.4.3
Werkzeug==1.0.1
xmlschema==1.9.2
yapf==0.29.0
zipp==3.6.0
